hatena_balloon assembled: 80/81 byte-exact, order matching
Not landed -- one function short, so the slice cannot be added yet. This
commits the assembled TU, the eight batch drafts, the agents' verifiers
and the header work, so none of it depends on the session surviving.

wip/hatena_balloon/hb_merged.cpp is 81 functions in target address order:
80 byte-identical, 18,216 B, emitted symbol order matching the target's
address order exactly. The gap is fly_ydisp_check -- 2 words of 55, two
adjacent independent lfs scheduled in the wrong order. Its agent swept
~85 variants across ten axes (accessor forms, cached pointer and
reference, 10 statement permutations, clamp spellings, operand binding
one side and both, signedness, early reads) without moving it.

Three gaps closed at assembly that no per-batch check could have caught,
each from a cross-batch disagreement:

- all_bgcheck returns int, not u8. u8 makes callers emit a clrlwi mask
  the target does not have. Two batches disagreed; the caller's author
  was right.
- dEnemyMng_c::m_110 is u32, not int. The code is `m_110--` then `> 3` --
  an unsigned wrap check -- and the target compares with cmplwi. Two
  batches disagreed here too, in the opposite direction.
- model_set's 7 words were not code at all. Our .data opened with
  "float_back"/"vibrate_back" because the animation-name arrays were
  declared at the top of the file; the target has model_set's own strings
  at g_profile+0xC. Moving the arrays next to their user fixed it.
  String-literal order in .data follows declaration position, so a
  file-scope array declared early displaces every later literal -- and it
  surfaces as wrong offsets in an unrelated function.

Header work this unit required, all verified against the bytes and all
symbol-neutral (CFront omits return types from mangling, so nothing but
the code can catch these): hipattackhit returns mVec3_c by value;
ccLineCheck returns u32; pause_check returns u8; five flight predicates
return bool; goalpole_check bool; floor_check u8; m_814 u32;
hitCallback_Ice is a tail call, not the bare blr the header claimed.
New: d_balloon_manager.hpp, plus declarations for cLib::chasePos,
envAllWaterCheck, floorEntryBufferCheck, mGoalPoleX and m_110. All five
binaries verify after every one of these edits.

+++ include/game/bases/d_a_en_hatena_balloon.hpp
@@ -172,7 +172,7 @@ public:
     /// mangle return types, so this changes no symbol. @unofficial
     bool goalpole_check();             ///< 0x80112580, 0x3C bytes.
     u8 floor_check();                  ///< 0x801125C0, 0x214 bytes. Returns `u8`, not void. @unofficial
-    u8 all_bgcheck(u8 &result);        ///< 0x801127E0, 0x164 bytes. Returns `u8`, not void. Walks #s_someCheckData 4x2 and calls @ref pointBgCheck for each entry. @unofficial
+    int all_bgcheck(u8 &result);        ///< 0x801127E0, 0x164 bytes. Returns `u8`, not void. Walks #s_someCheckData 4x2 and calls @ref pointBgCheck for each entry. @unofficial
     void fly_yspeed_set();             ///< 0x80112950, 0x1AC bytes. @unofficial
     void fly_xspeed_set(bool a);       ///< 0x80112B00, 0x164 bytes. @unofficial
     bool fly_ydisp_check(bool a);      ///< 0x80112C70, 0xDC bytes. @unofficial
